+++ previous-versions/common/src/codatcommon/models/operations/list_api_keys.py
@@ -3,27 +3,18 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import connection as shared_connection
+from ..shared import apikeys as shared_apikeys
 from ..shared import errormessage as shared_errormessage
 from typing import Optional
 
 
 
 @dataclasses.dataclass
-class GetCompanyConnectionRequest:
-    company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
-    connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
-    
-
-
-
-
-@dataclasses.dataclass
-class GetCompanyConnectionResponse:
+class ListAPIKeysResponse:
     content_type: str = dataclasses.field()
     status_code: int = dataclasses.field()
-    connection: Optional[shared_connection.Connection] = dataclasses.field(default=None)
-    r"""OK"""
+    api_keys: Optional[shared_apikeys.APIKeys] = dataclasses.field(default=None)
+    r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)