+++ previous-versions/common/src/codatcommon/models/shared/errormessage.py
@@ -11,7 +11,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class ErrorMessage:
-    r"""Your `query` parameter was not correctly formed"""
+    r"""Your API request was not properly authorized."""
     can_be_retried: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('canBeRetried'), 'exclude': lambda f: f is None }})
     correlation_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('correlationId'), 'exclude': lambda f: f is None }})
     detailed_error_code: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('detailedErrorCode'), 'exclude': lambda f: f is None }})