+++ previous-versions/common/RELEASES.md
@@ -591,3 +591,13 @@ Based on:
 - Speakeasy CLI 1.59.0 (2.65.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [PyPI v0.33.1] https://pypi.org/project/codat-common/0.33.1 - common
+
+## 2023-09-01 16:59:16
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.77.1 (2.91.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/common
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-common/0.1.0 - previous-versions/common